% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{BedtoolsIntersectQuery}
\alias{BedtoolsIntersectQuery}
\title{BedtoolsIntersectQuery}
\usage{
BedtoolsIntersectQuery(query = NULL, db = NULL, cutoff = 0.5)
}
\arguments{
\item{query}{query regions}

\item{db}{bed data}

\item{cutoff}{cutoff overlap}
}
\value{
data frame
}
\description{
BedtoolsIntersectQuery
}
