% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalCellRatioForMotifSig}
\alias{CalCellRatioForMotifSig}
\title{Calculate cell count/ratio for motif signal}
\usage{
CalCellRatioForMotifSig(obj, motifs = "all", group = "", subgroup = "")
}
\arguments{
\item{group}{name}

\item{subgroup}{name}

\item{object}{anno}

\item{motif}{ids}

\item{split}{by sample name yes/no}
}
\value{
data frame
}
\description{
Calculate cell count/ratio for motif signal
}
