% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalCellRatioForMotifSig_test}
\alias{CalCellRatioForMotifSig_test}
\title{Test Calculate cell count/ratio for motif signal}
\usage{
CalCellRatioForMotifSig_test(
  obj,
  motifs = "all",
  group = "",
  sample_group = "",
  cutoff = 0
)
}
\arguments{
\item{group}{name}

\item{object}{anno}

\item{motif}{ids}

\item{subgroup}{name}
}
\value{
data frame
}
\description{
Test Calculate cell count/ratio for motif signal
}
