% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{CalIntersectMotifOpenPeaksPerCell}
\alias{CalIntersectMotifOpenPeaksPerCell}
\title{CalIntersectMotifOpenPeaksPerCell}
\usage{
CalIntersectMotifOpenPeaksPerCell(
  obj = NULL,
  motif = "",
  fdb = "",
  cutoff = 0.5,
  count_cutoff = 2,
  temp_dir = "temp"
)
}
\arguments{
\item{obj}{multiome object}

\item{fdb}{bed file}

\item{cutoff}{overlap cutoff}

\item{count_cutoff}{count cutoff}

\item{temp_dir}{temporary dir}

\item{fquery}{bed file}
}
\value{
data frame
}
\description{
CalIntersectMotifOpenPeaksPerCell
}
