% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{CalIntersectOpenPeaksPerCell}
\alias{CalIntersectOpenPeaksPerCell}
\title{CalIntersectOpenPeaksPerCell}
\usage{
CalIntersectOpenPeaksPerCell(
  obj = NULL,
  assay = "peaks",
  motif = "",
  fdb = "",
  cutoff = 0.5,
  temp_dir = "temp"
)
}
\arguments{
\item{obj}{multiome object}

\item{fdb}{bed file}

\item{cutoff}{cutoff overlap}

\item{temp_dir}{temporary dir}

\item{fquery}{bed file}
}
\value{
data frame
}
\description{
CalIntersectOpenPeaksPerCell
}
