% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{CalIntersectPeaks}
\alias{CalIntersectPeaks}
\title{CalIntersectPeaks}
\usage{
CalIntersectPeaks(peaks = NULL, fbed = "", cutoff = 0.5, temp_dir = "temp")
}
\arguments{
\item{peaks}{bed file}

\item{fbed}{bed file}

\item{cutoff}{overlap cutoff}

\item{temp_dir}{temporary dir}
}
\value{
data frame
}
\description{
CalIntersectPeaks
}
