% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalMeanByGroup}
\alias{CalMeanByGroup}
\title{Call mean}
\usage{
CalMeanByGroup(df, group = "", with_sample = FALSE)
}
\arguments{
\item{df}{data}

\item{group}{name}
}
\value{
data frame
}
\description{
Call mean
}
