% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalMeanMotifSig}
\alias{CalMeanMotifSig}
\title{Calculate mean motif signal}
\usage{
CalMeanMotifSig(obj, motifs = "all", group = "", split = FALSE)
}
\arguments{
\item{group}{name}

\item{split}{by sample name yes/no}

\item{object}{anno}

\item{motif}{ids}
}
\value{
data frame
}
\description{
Calculate mean motif signal
}
