% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalMedianByGroup}
\alias{CalMedianByGroup}
\title{Call median}
\usage{
CalMedianByGroup(df, group = "", with_sample = FALSE)
}
\arguments{
\item{df}{data}

\item{group}{name}
}
\value{
data frame
}
\description{
Call median
}
