% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalSignalCountByGroup}
\alias{CalSignalCountByGroup}
\title{Call count/ratio}
\usage{
CalSignalCountByGroup(df, columns, group = "", subgroup = "", cutoff = 0)
}
\arguments{
\item{df}{data}

\item{group}{name}

\item{cutoff}{signal cutoff}
}
\value{
data frame
}
\description{
Call count/ratio
}
