% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{CC_SignalUMAP_perSample}
\alias{CC_SignalUMAP_perSample}
\title{CC_SignalUMAP_perSample}
\usage{
CC_SignalUMAP_perSample(
  df,
  sorted_sample = "",
  label_name = "Cell-Cycle\\nSignal"
)
}
\arguments{
\item{df}{dataframe}
}
\description{
CC_SignalUMAP_perSample
}
