% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{CC_UMAPPlot_Signal}
\alias{CC_UMAPPlot_Signal}
\title{CC_UMAPPlot_Signal}
\usage{
CC_UMAPPlot_Signal(df = NULL, label_name = "Cell-Cycle\\nSignal")
}
\arguments{
\item{df}{dataframe}
}
\description{
CC_UMAPPlot_Signal
}
