% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ElbowPlot2}
\alias{ElbowPlot2}
\title{Elbow Plot}
\usage{
ElbowPlot2(obj, reduc = "pca")
}
\arguments{
\item{obj}{object}

\item{reduc}{name}
}
\description{
Elbow Plot
}
\examples{
library(Seurat)
p <- ElbowPlot2(seu_obj = pbmc_small, reduc='pca')

}
