% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{EstimatedSignal}
\alias{EstimatedSignal}
\title{EstimatedSignal}
\usage{
EstimatedSignal(
  obj = NULL,
  assay_use = "SCT",
  name = "",
  features = "",
  min_cutoff = 0.1,
  max_cutoff = 0.2,
  umap = "umap"
)
}
\arguments{
\item{obj}{seurat object}

\item{features}{genes}

\item{umap}{umap}
}
\value{
data frame
}
\description{
EstimatedSignal
}
