% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ExportUMAP}
\alias{ExportUMAP}
\title{Export UMAP}
\usage{
ExportUMAP(object, umap)
}
\arguments{
\item{reference}{anno}
}
\value{
anotation
}
\description{
Export UMAP
}
\examples{
d_umap <- ExportUMAP(seu, 'wnn.umap')

}
