% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{ExtractMotifSig}
\alias{ExtractMotifSig}
\title{Extract motif signal}
\usage{
ExtractMotifSig(obj, motifs = "all")
}
\arguments{
\item{obj}{data}

\item{motifs}{ids}
}
\value{
data frame
}
\description{
Extract motif signal
}
