% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{FragmentDensityForGene}
\alias{FragmentDensityForGene}
\title{FragmentDensityForGene}
\usage{
FragmentDensityForGene(
  obj = NULL,
  fragment = "atac_fragments.tsv.gz",
  peaks = "peaks",
  gene = NULL,
  colors = c(RGC = "#BC243C", Neuron = "#0F4C81")
)
}
\arguments{
\item{obj}{seurat object}

\item{fragment}{fragment file}

\item{peaks}{peak name}

\item{gene}{gene name}

\item{colors}{color set}
}
\value{
data frame
}
\description{
FragmentDensityForGene
}
