% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenxConservedCellTypeMarkers}
\alias{GenxConservedCellTypeMarkers}
\title{Genx Conserved CellType Markers}
\usage{
GenxConservedCellTypeMarkers(obj, assay_use = "SCT", group = "cell_type2")
}
\arguments{
\item{obj}{object}

\item{assay_use}{assay}
}
\value{
data frame
}
\description{
Genx Conserved CellType Markers
}
