% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenxFindAllMarkers}
\alias{GenxFindAllMarkers}
\title{Gene expression markers for all group}
\usage{
GenxFindAllMarkers(
  obj,
  assay_use = "SCT",
  idents = "seurat_clusters",
  logfc = 0.25,
  min_pct = 0.05,
  min_diff_pct = 0,
  recorrect_umi = TRUE
)
}
\arguments{
\item{obj}{object}

\item{assay_use}{assay}
}
\value{
data frame
}
\description{
Gene expression markers for all group
}
