% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_FindAllMarkers}
\alias{Motif_FindAllMarkers}
\title{Motif find all markers}
\usage{
Motif_FindAllMarkers(obj)
}
\arguments{
\item{obj}{object}
}
\value{
data frame
}
\description{
Motif find all markers
}
