% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_FindAllMarkersMean}
\alias{Motif_FindAllMarkersMean}
\title{Motif find all markers Mean}
\usage{
Motif_FindAllMarkersMean(obj)
}
\arguments{
\item{obj}{object}
}
\value{
data frame
}
\description{
Motif find all markers Mean
}
