% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_FindMarkers}
\alias{Motif_FindMarkers}
\title{Motif find markers}
\usage{
Motif_FindMarkers(obj, query = "")
}
\arguments{
\item{obj}{object}

\item{query}{name}
}
\value{
data frame
}
\description{
Motif find markers
}
