% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_FindMarkersMean}
\alias{Motif_FindMarkersMean}
\title{Motif find markers Mean}
\usage{
Motif_FindMarkersMean(obj, query = "")
}
\arguments{
\item{obj}{object}

\item{query}{name}
}
\value{
data frame
}
\description{
Motif find markers Mean
}
