% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_SetSignalGroup}
\alias{Motif_SetSignalGroup}
\title{Motif SetSignalGroup}
\usage{
Motif_SetSignalGroup(
  obj,
  motif = NULL,
  group = "cell_type2",
  min_cutoff = 0.5,
  max_cutoff = 1
)
}
\arguments{
\item{obj}{object}

\item{motif}{id}

\item{group}{name}

\item{min_cutoff}{min signal cutoff}

\item{max_cutoff}{max signal cutoff}
}
\value{
data frame
}
\description{
Motif SetSignalGroup
}
