% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunFindConservedCellTypeMarkers}
\alias{RunFindConservedCellTypeMarkers}
\title{Find Conserved CellType Markers}
\usage{
RunFindConservedCellTypeMarkers(obj, assay = "SCT", group = "cell_type2")
}
\arguments{
\item{obj}{object}

\item{assay}{assay}
}
\value{
data frame
}
\description{
Find Conserved CellType Markers
}
