% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunFindMarkers}
\alias{RunFindMarkers}
\title{Find markers for two group}
\usage{
RunFindMarkers(
  obj = NULL,
  ct_target = NULL,
  ct_bg = NULL,
  recor_umi = TRUE,
  outdir = "."
)
}
\arguments{
\item{obj}{object}

\item{assay_use}{assay}
}
\value{
data frame
}
\description{
Find markers for two group
}
