% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scCustom_DimPlot}
\alias{scCustom_DimPlot}
\title{scCustom DimPlot}
\usage{
scCustom_DimPlot(
  obj = NULL,
  reduction = "umap",
  group_by = "cell_type2",
  title = "",
  x_lab = "UMAP 1",
  y_lab = "UMAP 2",
  colors = NULL,
  label = FALSE,
  repel = TRUE,
  figure_plot = TRUE
)
}
\arguments{
\item{obj}{object}

\item{reduction}{}
}
\value{
plot
}
\description{
scCustom DimPlot
}
