% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scCustom_FeaturePlot}
\alias{scCustom_FeaturePlot}
\title{scCustom FeaturePlot}
\usage{
scCustom_FeaturePlot(
  obj = NULL,
  slot = "data",
  gene = NULL,
  features = NULL,
  reduction = "wnn.umap",
  title = "",
  x_label = "UMAP 1",
  y_label = "UMAP 2",
  colors = viridis_plasma_dark_high
)
}
\arguments{
\item{obj}{object}

\item{features}{gene name}
}
\value{
plot
}
\description{
#gene exp: colors_use = viridis_inferno_dark_high
#gene act: colors_use = viridis_light_high
#motif: colors_use = viridis_plasma_dark_high
}
