% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scCustomize_ClusteredDotPlot}
\alias{scCustomize_ClusteredDotPlot}
\title{scCustomize ClusteredDotPlot}
\usage{
scCustomize_ClusteredDotPlot(obj = NULL, group = "", d_markers = "", n = 7)
}
\arguments{
\item{obj}{object}

\item{group}{}
}
\value{
plot
}
\description{
scCustomize ClusteredDotPlot
}
