% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scCustomize_DotPlotMarkers}
\alias{scCustomize_DotPlotMarkers}
\title{scCustomize DotPlotMarkers}
\usage{
scCustomize_DotPlotMarkers(obj = NULL, group = "", d_markers = "", n = 5)
}
\arguments{
\item{obj}{object}

\item{group}{}
}
\value{
plot
}
\description{
scCustomize DotPlotMarkers
}
