% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Seurat_DimPlot2}
\alias{Seurat_DimPlot2}
\title{Seurat DimPlot with arrow}
\usage{
Seurat_DimPlot2(
  obj = NULL,
  reduction = "umap",
  group_by = "cell_type2",
  title = "",
  x_lab = "UMAP 1",
  y_lab = "UMAP 2",
  colors = "",
  label = FALSE,
  repel = TRUE
)
}
\arguments{
\item{obj}{object}

\item{reduction}{}
}
\value{
plot
}
\description{
Seurat DimPlot with arrow
}
