% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlot.R
\name{Seurat_DotPlot}
\alias{Seurat_DotPlot}
\title{Seurat DotPlot - default motif}
\usage{
Seurat_DotPlot(
  obj = NULL,
  scale = FALSE,
  assay = "chromvar",
  features = NULL,
  flip = FALSE,
  col_min = 0,
  cex = 6,
  lws = 0.3,
  title = "",
  title_size = 7,
  text_size = 6,
  col_set = "motif"
)
}
\arguments{
\item{obj}{object}

\item{assay}{assay}
}
\value{
plot
}
\description{
Seurat DotPlot - default motif
}
