% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlot.R
\name{Seurat_DotPlotMarkers}
\alias{Seurat_DotPlotMarkers}
\title{Seurat DotPlot}
\usage{
Seurat_DotPlotMarkers(
  obj = NULL,
  assay = "SCT",
  d_markers = NULL,
  n = 5,
  group = ""
)
}
\arguments{
\item{obj}{object}

\item{group}{}
}
\value{
plot
}
\description{
Seurat DotPlot
}
