% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Seurat_FeaturePlot}
\alias{Seurat_FeaturePlot}
\title{Seurat FeaturePlot}
\usage{
Seurat_FeaturePlot(
  obj = NULL,
  slot = "data",
  reduction = "umap",
  features = NULL,
  colorset = "lightgrey,red",
  xlab = "UMAP 1",
  ylab = "UMAP 2"
)
}
\arguments{
\item{seurat}{object}

\item{feature}{name}
}
\value{
data frame
}
\description{
Seurat FeaturePlot
}
