% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SeuratAddScore}
\alias{SeuratAddScore}
\title{Seurat Add Score}
\usage{
SeuratAddScore(
  obj = NULL,
  assay = "SCT",
  slot = "data",
  name = "",
  features = ""
)
}
\arguments{
\item{obj}{seurat object}

\item{name}{score title}

\item{features}{gene vectors}
}
\value{
seurat object
}
\description{
Seurat Add Score
}
