% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{SeuratCellCycleScoring}
\alias{SeuratCellCycleScoring}
\title{Seurat_EstimateCycling}
\usage{
SeuratCellCycleScoring(
  obj = NULL,
  ref = "mouse",
  s.genes = NULL,
  g2m.genes = NULL,
  umap = "umap"
)
}
\arguments{
\item{obj}{seurat object}

\item{ref}{reference}

\item{s.genes}{gene vectors}

\item{g2m.genes}{gene vectors}
}
\value{
data frame
}
\description{
Seurat_EstimateCycling
}
