% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Signal_UMAPPlot}
\alias{Signal_UMAPPlot}
\title{Signal UMAPPlot Group}
\usage{
Signal_UMAPPlot(
  df = NULL,
  label_name = "Signal",
  breaks = c("High", "Middle", "Low"),
  colors = c("#CB4335", "#F5B7B1", "#D7DBDD")
)
}
\arguments{
\item{df}{dataframe}
}
\description{
Signal UMAPPlot Group
}
