% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalCellRatioForMotifSig}
\alias{CalCellRatioForMotifSig}
\title{Calculate cell count/ratio for motif signal}
\usage{
CalCellRatioForMotifSig(
  obj,
  motifs = "all",
  group = "",
  subgroup = "",
  split = FALSE
)
}
\arguments{
\item{group}{name}

\item{subgroup}{name}

\item{split}{by sample name yes/no}

\item{object}{anno}

\item{motif}{ids}
}
\value{
data frame
}
\description{
Calculate cell count/ratio for motif signal
}
