% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalCellRatioForMotifSig}
\alias{CalCellRatioForMotifSig}
\title{Calculate cell count/ratio for motif signal}
\usage{
CalCellRatioForMotifSig(obj, motifs = "all", g1 = "", g2 = "", split = FALSE)
}
\arguments{
\item{split}{by sample name yes/no}

\item{object}{anno}

\item{motif}{ids}

\item{group}{name}
}
\value{
data frame
}
\description{
Calculate cell count/ratio for motif signal
}
