% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalCellRatioForMotifSig}
\alias{CalCellRatioForMotifSig}
\title{Calculate cell count/ratio for motif signal}
\usage{
CalCellRatioForMotifSig(
  obj,
  motifs = NULL,
  group = "",
  sample_group = "",
  cutoff = 0
)
}
\arguments{
\item{obj}{anno}

\item{group}{name}

\item{sample_group}{name}

\item{motif}{ids}
}
\value{
data frame
}
\description{
Calculate cell count/ratio for motif signal
}
