% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CallSignalCountRatioByGroup}
\alias{CallSignalCountRatioByGroup}
\title{Call count/ratio}
\usage{
CallSignalCountRatioByGroup(df, columns, g1 = "", g2 = "", cutoff = 0)
}
\arguments{
\item{df}{data}

\item{cutoff}{signal cutoff}

\item{group}{name}
}
\value{
data frame
}
\description{
Call count/ratio
}
