% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalMotifMeanOrMedianSig}
\alias{CalMotifMeanOrMedianSig}
\title{Calculate mean motif signal}
\usage{
CalMotifMeanOrMedianSig(obj, motifs = "all", group = "", method = "mean")
}
\arguments{
\item{obj}{anno}

\item{group}{name}

\item{motif}{ids}
}
\value{
data frame
}
\description{
Calculate mean motif signal
}
