% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CalMotifMeanSig}
\alias{CalMotifMeanSig}
\title{Calculate mean motif signal}
\usage{
CalMotifMeanSig(obj, motifs = "all", group = "")
}
\arguments{
\item{obj}{anno}

\item{group}{name}

\item{motif}{ids}
}
\value{
data frame
}
\description{
Calculate mean motif signal
}
