% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{ExtractMotifSigTranspose}
\alias{ExtractMotifSigTranspose}
\title{Extract multiple motif set signal}
\usage{
ExtractMotifSigTranspose(obj, motifs = NULL, group = "")
}
\arguments{
\item{obj}{data}

\item{motifs}{ids}

\item{group}{add group name}
}
\value{
data frame
}
\description{
Extract multiple motif set signal
}
