% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{Motif_FindMarkers}
\alias{Motif_FindMarkers}
\title{Motif find markers}
\usage{
Motif_FindMarkers(
  obj,
  query = "",
  avg_fc = 1,
  cutoff_pct1 = 0.2,
  cutoff_diff_pct = 0.1
)
}
\arguments{
\item{object}{anno}

\item{motif}{ids}

\item{group}{name}

\item{subgroup}{name}
}
\value{
data frame
}
\description{
Motif find markers
}
