% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunFindAllMarkers}
\alias{RunFindAllMarkers}
\title{Find markers for all group}
\usage{
RunFindAllMarkers(
  obj = NULL,
  assay = "SCT",
  idents = "seurat_clusters",
  logfc = 0.25,
  min_pct = 0.1,
  min_diff_pct = 0,
  recor_umi = TRUE
)
}
\arguments{
\item{obj}{object}

\item{assay}{assay}
}
\value{
data frame
}
\description{
Find markers for all group
}
