% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RunFindMarkers}
\alias{RunFindMarkers}
\title{Find markers for two group}
\usage{
RunFindMarkers(
  obj = NULL,
  assay = "SCT",
  idents = "seurat_clusters",
  logfc = 0.25,
  only_pos = TRUE,
  ct_target = NULL,
  ct_bg = NULL,
  recor_umi = TRUE
)
}
\arguments{
\item{obj}{object}

\item{assay}{assay}
}
\value{
data frame
}
\description{
Find markers for two group
}
