% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlot.R
\name{Seurat_DotPlot}
\alias{Seurat_DotPlot}
\title{Seurat DotPlot - default motif}
\usage{
Seurat_DotPlot(
  obj = NULL,
  scale = FALSE,
  set_scale_size = TRUE,
  assay = "SCT",
  features = NULL,
  flip = FALSE,
  col_min = -2.5,
  col_max = 2.5,
  dot_min = 0,
  cex = 6,
  lws = 0.3,
  title = "",
  title_size = 7,
  text_size = 6,
  title_ht = "Average\\nExpression",
  title_dot = "Cell Percentage",
  col_set = "motif",
  col_direction = 1
)
}
\arguments{
\item{obj}{object}

\item{assay}{assay}
}
\value{
plot
}
\description{
Seurat DotPlot - default motif
}
