% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlot.R
\name{Seurat_DotPlot}
\alias{Seurat_DotPlot}
\title{Seurat DotPlot - default motif}
\usage{
Seurat_DotPlot(
  obj = NULL,
  assay_use = "chromvar",
  features = NULL,
  col_min = 0,
  cex = 6,
  colors = "viridis::mako"
)
}
\arguments{
\item{obj}{object}

\item{assay_use}{assay}
}
\value{
plot
}
\description{
Seurat DotPlot - default motif
}
