% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlot.R
\name{Seurat_FeaturePlot}
\alias{Seurat_FeaturePlot}
\title{Seurat FeaturePlot}
\usage{
Seurat_FeaturePlot(
  obj = NULL,
  slot = "data",
  reduction = "umap",
  features = NULL,
  colors = c("#E0E0E0", "red"),
  pt_size = 0.01,
  title = "",
  xlab = "UMAP 1",
  ylab = "UMAP 2",
  order = TRUE,
  min_cutoff = NA,
  max_cutoff = NA
)
}
\arguments{
\item{obj}{object}

\item{features}{gene name}
}
\value{
plot
}
\description{
Seurat FeaturePlot
}
