% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Seurat_FeaturePlot}
\alias{Seurat_FeaturePlot}
\title{Seurat FeaturePlot}
\usage{
Seurat_FeaturePlot(
  obj = NULL,
  slot = "data",
  reduction = "umap",
  features = NULL,
  colorset = "#E0E0E0,red",
  title = "",
  xlab = "UMAP 1",
  ylab = "UMAP 2",
  min_cutoff = NA,
  max_cutoff = NA
)
}
\arguments{
\item{obj}{object}

\item{features}{gene name}
}
\value{
data frame
}
\description{
Seurat FeaturePlot
}
